==> 02-scanpy_EDA.Rmd <==

# load data from seurat:
```{python eval=FALSE}
# conda activate scvi-env
import scanpy as sc
import anndata
from scipy import io
from scipy.sparse import coo_matrix, csr_matrix
import numpy as np
import seaborn as sns
from matplotlib import pyplot as plt
import os
import pandas as pd

data_dir = "data/"
fig_dir = "figures/"

# load sparse matrix:
X = io.mmread("{}FIRE_mouse_normCounts.mtx".format(data_dir))

# create anndata object
adata = anndata.AnnData(
    X=X.transpose().tocsr()
)

# load sample metadata:
sample_meta = pd.read_csv("{}FIRE_mouse_meta.csv".format(data_dir))

# load gene names:
with open("{}gene_names.csv".format(data_dir), 'r') as f:
    gene_names = f.read().splitlines()

adata.obs = sample_meta
adata.obs.index = adata.obs['barcode']
adata.obs = adata.obs.drop(labels='barcode', axis=1)
adata.var.index = gene_names

# load pca:
pca = pd.read_csv("{}FIRE_mouse_harmony.csv".format(data_dir))
pca.index = adata.obs.index

# set pca and umap
adata.obsm['X_pca'] = pca.to_numpy()
adata.obsm['X_umap'] = np.vstack((adata.obs['UMAP1'].to_numpy(), adata.obs['UMAP2'].to_numpy())).T

# compute umap with scanpy:
sc.pp.neighbors(adata, n_neighbors=15, n_pcs=30, metric='cosine')
sc.tl.umap(adata, min_dist=0.1)
sc.pl.umap(adata, color='DX', save='_scanpy.pdf', frameon=False)

# correct with PAGA:
sc.pp.neighbors(adata, n_neighbors=15, n_pcs=30, metric='cosine')
sc.tl.leiden(adata, resolution=1.5)
adata.obs['leiden_res1.5'] = adata.obs['leiden']
sc.tl.paga(adata)
sc.pl.paga(
    adata, threshold=0.3, title='', edge_width_scale=0.5,
    frameon=False, save=True
)

# re-compute UMAP
sc.tl.umap(adata, min_dist=0.15, init_pos='paga')
sc.pl.umap(adata, color='leiden', save='_paga_leiden.pdf', frameon=False)

# compute low-res leiden clusters
sc.tl.leiden(adata, resolution=0.7)
adata.obs['leiden_res0.7'] = adata.obs['leiden']

# compare scanpy clusters with seurat clusters:
sc.pl.umap(adata, color=['clusternum_anno', 'leiden_res0.7', 'leiden_res1.5'], save='_paga_compare_clusters.pdf', frameon=False, legend_loc='on data')

# save anndata object
adata.write('{}FIRE_mouse_scanpy.h5ad'.format(data_dir))

# save cluster assignments and UMAP coordinates for Seurat
adata.obs['paga_UMAP1'] = adata.obsm['X_umap'][:,0]
adata.obs['paga_UMAP2'] = adata.obsm['X_umap'][:,1]
adata.obs.to_csv('{}scanpy_meta.csv'.format(data_dir))


```
